requirements.txt
praw
pandas
numpy
matplotlib
seaborn
kafka-python
flask